requirements.txt
# Core dependencies
ccxt>=4.0.0
websockets>=12.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Async support
asyncio
aiohttp>=3.9.0

# Logging and monitoring
loguru>=0.7.0

# Data validation
pydantic>=2.0.0

# Time and timezone handling (critical for trading hours)
pytz>=2024.1
arrow>=1.3.0

# API communication enhancements
httpx>=0.25.0  # Modern async HTTP client

# Caching for market data
cachetools>=5.3.0

# Configuration formats
pyyaml>=6.0  # For config files

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
pytest-xdist>=3.0.0  # Parallel test execution

# Testing utilities
faker>=18.0.0  # Generate fake data
freezegun>=1.2.0  # Mock datetime
responses>=0.23.0  # Mock HTTP requests

# Future enhancements (commented out for now)
# rich>=13.0.0  # Terminal dashboard